+++ packages/tinkoff_acquiring/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.0.2
+
+* Add `getQr` endpoint for SBP
+* Add `getStaticQr` endpoint for SBP
+
 # 3.0.1
 
 * Add SendClosingReceipt endpoint